Move CheckboxListItem to ui/widgets
