+++ picard/ui/options/cover.py
@@ -38,11 +38,11 @@ from picard.i18n import (
     gettext as _,
 )
 
-from picard.ui.checkbox_list_item import CheckboxListItem
 from picard.ui.forms.ui_options_cover import Ui_CoverOptionsPage
 from picard.ui.moveable_list_view import MoveableListView
 from picard.ui.options import OptionsPage
 from picard.ui.util import qlistwidget_items
+from picard.ui.widgets.checkbox_list_item import CheckboxListItem
 
 
 class CoverOptionsPage(OptionsPage):
