+++ picard/ui/widgets/checkbox_list_item.py
